move smoke/dust ush scripts to common directory
