+++ ush/smoke_dust/fire_emiss_tools.py
