+++ ush/smoke_dust/generate_fire_emissions.py
@@ -11,7 +11,7 @@
 import sys
 import os
 import fire_emiss_tools as femmi_tools
-import HWP_tools
+import hwp_tools
 import interp_tools as i_tools
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -95,9 +95,9 @@ def generate_emiss_workflow(staticdir: str, ravedir: str, intp_dir: str, predef_
                femmi_tools.produce_emiss_24hr_file(frp_avg_reshaped, nwges_dir, current_day, tgt_latt, tgt_lont, ebb_total_reshaped, cols, rows)
            elif ebb_dcycle == 2:              
                print('Restart dates to process',fcst_dates)
-               hwp_avail_hours, hwp_non_avail_hours = HWP_tools.check_restart_files(hourly_hwpdir, fcst_dates)
-               restart_avail, restart_nonavail_hours_test = HWP_tools.copy_missing_restart(nwges_dir, hwp_non_avail_hours, hourly_hwpdir, len_restart_interval)
-               hwp_ave_arr, xarr_hwp, totprcp_ave_arr, xarr_totprcp = HWP_tools.process_hwp(fcst_dates, hourly_hwpdir, cols, rows, intp_dir, rave_to_intp)
+               hwp_avail_hours, hwp_non_avail_hours = hwp_tools.check_restart_files(hourly_hwpdir, fcst_dates)
+               restart_avail, restart_nonavail_hours_test = hwp_tools.copy_missing_restart(nwges_dir, hwp_non_avail_hours, hourly_hwpdir, len_restart_interval)
+               hwp_ave_arr, xarr_hwp, totprcp_ave_arr, xarr_totprcp = hwp_tools.process_hwp(fcst_dates, hourly_hwpdir, cols, rows, intp_dir, rave_to_intp)
                frp_avg_reshaped, ebb_total_reshaped = femmi_tools.averaging_FRP(ebb_dcycle, fcst_dates, cols, rows, intp_dir, rave_to_intp, veg_map, tgt_area, beta, fg_to_ug, to_s)
                #Fire end hours processing
                te = femmi_tools.estimate_fire_duration(intp_dir, fcst_dates, current_day, cols, rows, rave_to_intp)
